manage jenkins location config
The Mesos plugin tends to throw NullPointer exceptions if the framework
registers but the Jenkins URL doesn't already exist within
jenkins.model.JenkinsLocationConfiguration.xml. This leads to a slave
trying to come online, but it never connects, times out, and destroys
itself.

This commit manages this config, and adds a sed statement to the
standalone script to populate the URL based on $HOST_IP and $PORT.